lyrics lines refactoring
