--- src/components/Editor/chordsEditor/Line.tsx
@@ -1,10 +1,0 @@
-import React from 'react';
-
-export interface LineProps {
-    text: string;
-    onEnter: () => void;
-}
-
-export const Line: React.FC = () => {
-    return <></>;
-};
