+++ src/components/Editor/chordsEditor/components/Line/EditLine.scss
