+++ src/components/Editor/chordsEditor/components/Line/LyricsLine/Letter.scss
