+++ src/components/Editor/chordsEditor/components/Line/LyricsLine/Letter.tsx
@@ -1,6 +1,6 @@
-import React, { Key } from 'react';
+import React from 'react';
 import { Tooltip, Whisper } from 'rsuite';
-import { ChordEditPopup } from '../ChordEditPopup';
+import { ChordEditPopup } from '../../ChordEditPopup';
 import cn from 'classnames';
 import { ChordType } from '@model/song';
 import './Letter.scss';
