+++ src/components/Editor/chordsEditor/components/Line/LyricsLine/LyricsLine.scss
