+++ src/components/Editor/chordsEditor/components/Line/LyricsLine/index.ts
