+++ src/components/Editor/chordsEditor/components/Line/index.ts
@@ -1,0 +1,1 @@
+export * from './Line';
